requirements.txt
aws-cdk-lib
constructs
autopep8
aws-cdk.core